Better Country, Currency Selection
Created list of all countries in the world, with currencies and currency
codes.
Choose Country in onboarding is now built from country model.
Choose Currency step added to onboarding, built from country model.
Added a Base View to load the user model and set language, this will be
used to hold "global" that needs to be consistent across views. In the
future, refactor the router to use methods inside of base to create and
destroy views.